requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
click==8.2.1; python_version >= '3.10'
fastapi==0.115.12; python_version >= '3.8'
h11==0.16.0; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
pydantic==2.11.7; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
python-dotenv==1.1.0; python_version >= '3.9'
python-json-logger==3.3.0; python_version >= '3.8'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.46.2; python_version >= '3.9'
typing-extensions==4.14.0; python_version >= '3.9'
typing-inspection==0.4.1; python_version >= '3.9'
uvicorn==0.34.3; python_version >= '3.9'
